+++ backend/de.metas.acct.base/src/main/java-legacy/org/compiere/acct/Doc.java
@@ -64,8 +64,8 @@ import de.metas.acct.api.AcctSchemaId;
 import de.metas.acct.doc.AcctDocContext;
 import de.metas.acct.doc.AcctDocRequiredServicesFacade;
 import de.metas.acct.doc.PostingException;
+import de.metas.banking.BankAccountId;
 import de.metas.banking.api.BPBankAccountAcct;
-import de.metas.banking.api.BankAccountId;
 import de.metas.bpartner.BPartnerId;
 import de.metas.currency.CurrencyConversionContext;
 import de.metas.currency.CurrencyPrecision;
