+++ backend/de.metas.acct.base/src/main/java-legacy/org/compiere/acct/Doc_Cash.java
@@ -30,7 +30,7 @@ import org.compiere.util.Env;
 import de.metas.acct.api.AcctSchema;
 import de.metas.acct.api.PostingType;
 import de.metas.acct.doc.AcctDocContext;
-import de.metas.banking.api.BankAccountId;
+import de.metas.banking.BankAccountId;
 import de.metas.money.CurrencyId;
 
 /**
