+++ backend/de.metas.acct.base/src/main/java/de/metas/acct/doc/AcctDocRequiredServicesFacade.java
@@ -29,9 +29,9 @@ import de.metas.acct.api.IProductAcctDAO;
 import de.metas.acct.api.ProductAcctType;
 import de.metas.acct.tax.ITaxAcctBL;
 import de.metas.acct.tax.TaxAcctType;
+import de.metas.banking.BankAccountId;
 import de.metas.banking.api.BPBankAccountAcct;
 import de.metas.banking.api.BPBankAccountAcctRepository;
-import de.metas.banking.api.BankAccountId;
 import de.metas.banking.api.IBPBankAccountDAO;
 import de.metas.cache.model.CacheInvalidateMultiRequest;
 import de.metas.cache.model.IModelCacheInvalidationService;
