+++ backend/de.metas.banking/de.metas.banking.base/src/main/java-legacy/org/compiere/acct/Doc_BankStatement.java
@@ -30,9 +30,9 @@ import com.google.common.collect.ImmutableList;
 import de.metas.acct.api.AcctSchema;
 import de.metas.acct.api.PostingType;
 import de.metas.acct.doc.AcctDocContext;
+import de.metas.banking.BankAccountId;
 import de.metas.banking.BankStatementId;
 import de.metas.banking.BankStatementLineReference;
-import de.metas.banking.api.BankAccountId;
 import de.metas.banking.service.IBankStatementBL;
 import de.metas.bpartner.BPartnerId;
 import de.metas.currency.CurrencyConversionContext;
