+++ backend/de.metas.banking/de.metas.banking.base/src/main/java/de/metas/banking/payment/impl/BankStatementPaymentBL.java
@@ -22,9 +22,9 @@
 
 package de.metas.banking.payment.impl;
 
+import de.metas.banking.BankAccountId;
 import de.metas.banking.BankStatementId;
 import de.metas.banking.BankStatementLineId;
-import de.metas.banking.api.BankAccountId;
 import de.metas.banking.payment.BankStatementLineMultiPaymentLinkRequest;
 import de.metas.banking.payment.BankStatementLineMultiPaymentLinkResult;
 import de.metas.banking.payment.IBankStatementPaymentBL;
