+++ backend/de.metas.banking/de.metas.banking.base/src/main/java/de/metas/banking/service/IBankStatementDAO.java
@@ -14,12 +14,12 @@ import org.compiere.model.I_C_Payment;
 
 import com.google.common.collect.ImmutableSet;
 
+import de.metas.banking.BankAccountId;
 import de.metas.banking.BankStatementId;
 import de.metas.banking.BankStatementLineId;
 import de.metas.banking.BankStatementLineRefId;
 import de.metas.banking.BankStatementLineReference;
 import de.metas.banking.BankStatementLineReferenceList;
-import de.metas.banking.api.BankAccountId;
 import de.metas.document.engine.DocStatus;
 import de.metas.payment.PaymentId;
 import de.metas.util.ISingletonService;
