+++ backend/de.metas.banking/de.metas.banking.base/src/main/java/de/metas/banking/service/impl/CashStatementBL.java
@@ -12,9 +12,9 @@ import org.compiere.model.I_C_Payment;
 import org.compiere.model.Query;
 import org.compiere.util.TimeUtil;
 
+import de.metas.banking.BankAccountId;
 import de.metas.banking.BankStatementId;
 import de.metas.banking.BankStatementLineId;
-import de.metas.banking.api.BankAccountId;
 import de.metas.banking.payment.IBankStatementPaymentBL;
 import de.metas.banking.service.BankStatementCreateRequest;
 import de.metas.banking.service.BankStatementLineCreateRequest;
